Rename service.py to server.py
